Refactor: images 폴더 이동 및 경로 변경
